+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 * Changed the import behavior for wismhd maps to use collection instances instead of linked duplicate meshes.
 * Changed exception message formatting to include the inner errors for easier debugging.
+* Moved normal map Z calculation to a node group.
 
 ## 0.8.2 - 2024-06-12
 ### Added